better submodule and update
