+++ CHANGES.txt
@@ -1,3 +1,9 @@
+1.1.0 (2020-05-13)
+------------------
+- better submodule definition
+- add `morecantile.tms` object to access and register defaults TileMatrixSet
+- Add depreciation warning for `morecantile.TileMatrixSet.load` method
+
 1.0.0 (2020-05-11)
 ------------------
 Major refactor of Morecantile, which is now based on OGC TileMatrixSet JSON documents.
